+++ deps.sh
@@ -6,6 +6,7 @@ set -e
 # Test dependencies:
 tarantoolctl rocks install luatest 0.5.3
 tarantoolctl rocks install luacheck 0.25.0
+tarantoolctl rocks install luacov 0.13.0
 
 tarantoolctl rocks install cartridge 2.5.1
 tarantoolctl rocks make