+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* `crud.len()` function to iterate over each shard and count space size
+* `crud.len()` function to calculate the number of tuples
+  in the space for memtx engine and calculate the maximum
+  approximate number of tuples in the space for vinyl engine.
 
 ## [0.8.0] - 02-07-21